removed download list fragment, moved files to list package
ListTabFragment is the new list fragment
