+++ app/app/src/main/java/com/github/livingwithhippos/unchained/lists/model/DownloadPagingSource.kt
@@ -3,7 +3,6 @@ package com.github.livingwithhippos.unchained.lists.model
 import androidx.paging.PagingSource
 import com.github.livingwithhippos.unchained.base.model.repositories.CredentialsRepository
 import com.github.livingwithhippos.unchained.base.model.repositories.DownloadRepository
-import com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem
 import retrofit2.HttpException
 import java.io.IOException
