+++ app/app/src/main/java/com/github/livingwithhippos/unchained/lists/view/ListsTabFragment.kt
@@ -9,16 +9,12 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import androidx.paging.PagingData
-import com.github.livingwithhippos.unchained.R
 import com.github.livingwithhippos.unchained.base.UnchainedFragment
 import com.github.livingwithhippos.unchained.databinding.FragmentTabListsBinding
-import com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem
-import com.github.livingwithhippos.unchained.downloadlists.view.DownloadListFragmentDirections
-import com.github.livingwithhippos.unchained.downloadlists.view.DownloadListListener
-import com.github.livingwithhippos.unchained.downloadlists.view.DownloadListPagingAdapter
-import com.github.livingwithhippos.unchained.downloadlists.viewmodel.DownloadListViewModel
+import com.github.livingwithhippos.unchained.lists.model.DownloadItem
 import com.github.livingwithhippos.unchained.lists.model.TorrentListListener
 import com.github.livingwithhippos.unchained.lists.model.TorrentListPagingAdapter
+import com.github.livingwithhippos.unchained.lists.viewmodel.DownloadListViewModel
 import com.github.livingwithhippos.unchained.newdownload.model.TorrentItem
 import com.github.livingwithhippos.unchained.start.viewmodel.MainActivityViewModel
 import com.google.android.material.tabs.TabLayout
