+++ app/app/src/main/res/layout-land/fragment_download_details.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="details"
-            type="com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem" />
+            type="com.github.livingwithhippos.unchained.lists.model.DownloadItem" />
 
         <variable
             name="stream"
