+++ app/app/src/main/res/layout/item_list_download.xml
@@ -9,11 +9,11 @@
 
         <variable
             name="item"
-            type="com.github.livingwithhippos.unchained.downloadlists.model.DownloadItem" />
+            type="com.github.livingwithhippos.unchained.lists.model.DownloadItem" />
 
         <variable
             name="listener"
-            type="com.github.livingwithhippos.unchained.downloadlists.view.DownloadListListener" />
+            type="com.github.livingwithhippos.unchained.lists.view.DownloadListListener" />
     </data>
  <FrameLayout
      android:layout_width="match_parent"
